图片上传

+++ iseeu/ClassificationTableView.m
@@ -48,7 +48,7 @@
     
     _tableView.separatorStyle = UITableViewCellSeparatorStyleNone;
     
-    NSLog(@"%i",self.tag);
+    NSLog(@"%li",(long)self.tag);
     
     [self addSubview:_tableView];
